% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAdownload.R
\name{EPAdownload}
\alias{EPAdownload}
\title{A Self-made function for downloading hourly monitoring data from Taiwan EPA.}
\usage{
EPAdownload(start.time, end.time = NULL, epa.site = NULL, outputname = "EPA")
}
\arguments{
\item{start.time}{The starting time of data downloaded. The date format follows "YYYY-mm-dd HH:MM". Nevertheless, formats including "YYYY", "YYYY-mm", and "YYYY-mm-dd" are also feasible. Note that data is only available since July, 1982.}

\item{end.time}{The ending time of data downloaded. The date format follows "YYYY-mm-dd HH:MM". Nevertheless, formats including "YYYY", "YYYY-mm" are also feasible. The default is current local time.}

\item{epa.site}{One or more than one EPA site can be selected by inputting its Chinese name. The default is all sites.}

\item{outputname}{The name of the output file. The default is "EPA".}
}
\description{
A preferred time interval can be selected by defining the starting and ending time. The function setwd() is recommended to use beforhand to specify the path to the preferred folder.
}
