% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSensorsPlot.R
\name{MultiSensorsPlot}
\alias{MultiSensorsPlot}
\title{A function for plotting with one variable (multiple sensors) per plot.}
\usage{
MultiSensorsPlot(
  data_path,
  imeiID,
  sitenames,
  mode = "both",
  x_breaks = "2 days",
  x_labels = "\%d",
  N_cols = 2,
  font = "STKaiti",
  width = 40,
  height = 30
)
}
\arguments{
\item{data_path}{Required. The path of the raw "ens" data downloaded.}

\item{imeiID}{Required. Note that the order of the imeiID input should be paired with the order of sitenames input.}

\item{sitenames}{Required. Note that the order of the sitenames input should be paired with the order of imeiID input.}

\item{mode}{The output could be either time series("ts"), hourly average("havg"), or both time series and hourly average plots("both"). The default is "both".}

\item{x_breaks}{It define the X axis intervals for time series plots only. The format of x_breaks input could be, for example, "1 hour" , "12 hours" , "1 day" , "2 weeks", and etc. The default is 2 days. For hourly average plots, the X axis intervals is fixed to 2 hours.}

\item{x_labels}{It define the X axis labels for time series plots only. The format of x_breaks input could be, for example, \code{"\%d \%H"} , \code{"\%m/\%d \%H"} , \code{"\%Y/\%m/\%d"} , and etc. The default is \code{"\%H"} .}

\item{N_cols}{The number of the columns of the plot. The default is 1.}

\item{font}{The font family of the plot. The default is "STKaiti". Please be awared that the default is for iOS users. For windows users, please change the font to, for example, "KaiTi".}

\item{width}{The width of the plot output. The default is 40 cm.}

\item{height}{The height of the plot output. The default is 30 cm.}
}
\description{
It is used for exporting multiple graphs with one variable (multiple sensors) on each graph, and a total of five variables("pm25" , "voc" , "co2" , "temperature" , "humidity") is included. It is particularly used for the sensors located at the nursing centers. The raw "ens" data is applied to the function directly. The function setwd() is recommended to use beforhand to specify the file path of these plots to a preferred folder.
}
